requirements.txt
discord.py[voice]
yt-dlp
youtube-search-python
asyncio
shutil
python-dotenv
importlib
glob
